% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expoviz.R
\name{FuncExit}
\alias{FuncExit}
\title{End the module analysis}
\usage{
FuncExit(PID)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitViz}
}
\description{
End the module analysis
}
\examples{
OutPath = getwd()
   res = InitViz()
   res = LoadViz(PID = res$PID,UseExample = "example#1")
   res2 = VizCateDot(PID=res$PID,OutPath = Path,Group = "T",Vars = "X4,X5,X6,X7,X8,X9,X10",Parameter = "mean",
   Brightness = "light",Palette = "default1")
   FuncExit(PID = res$PID)
}
\author{
Ning Gao,Bin Wang(corresponding author)
}
